Ignore whitespace after parameter delimeter
When a user entered a line with whitespace after the parameter
delimeter(;), that whitespace was included in the key for that key/value
pair. For example, if "* * * * * % key1=value1; key2=value2" was
entered, the second key to be " key2" rather than "key2". Invoking
`trimResults()` on the splitter resolves the issue.